requirements.txt
flask
slack_sdk
python-dotenv
google-generativeai
